+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * New `app component cron` commands for the new crons feature.
 * `lvl login` now accepts username from command via `-u`, and can correctly read the password when piped via stdin. This allows automating it.
 * `lvl update` allows you to easily update `lvl` to the latest version.
+* Fix API marshalling errors with system checks and app component types.
 
 ## 1.6.0